Fix minor bugs & add design resources
A bug on the site footer meant it could not 'hide' and 'show' per user
click on the interface. The bug was rectified on the JQuery controller
file (tut.js). Consequently, since the system controller had been passed
to a central document, the onclick call to doclose on all pages across
the site becomes a redundant call and can be taken off.

4 images were included to build the Site Attendant. The images represent
Attendant opening, closing, waiting and working.