Introduce Site Assitant
Test a new tool on the Tools menu. The Site Assistant is a GUI
alternative for users to search through content on the site and make
inquiries into topics of their choice, ranging from Cryptocurrency to
Forex, CFD and even Blockchain topics.
Minor changes are made to layout and bug fixes are also made.