Fix minor bugs
On page resizing, footer contents were seen to lose their properties.
The JQuery controls for setting size of layout depends on user click.
The control for size regulation on the pages was passed directly to CSS.
Only controls for layout properties of expandable columns on 1-col
layouts are handled with JQuery when the user may click and close the
floating column.