Fix minor bugs
The Back and Next buttons on the menu were not centralized. Hence, on
dropdown views, the buttons float to the left. It was important that the
buttons on the dropdown list maintain consistency. Due to this, the both
buttons were also centralized.
The Site Assistant visible on the binary-options-guide-eu page, had a
looping gif image. It was considered unnecessary that the image
continues to loop, hence, the image was converted into an Animated GIF
and its timeline was set to run once and freeze.