Restart component update
Reset available components and their styling. The About.vue component is a
2-column layout, whilst the Home.vue component is a 3-column layout.
Favicon is set as Tutusfx logo and dependencies are reduced to include
only material design icon, material icon, router, vuetify and vuex.
The application is built using Vue-CLI hence, babel and eslint come as
default dependencies in the application.